applied review from sonots, remove Core::FluentLoggerBase and logger interface implemented.
